fix: extract ingested entity from proto
- reason: we are using protobuf data instead of decoding with mapstructure (used mainly for json [stdlib] marshalled structs)
- add methods for converting protobuf entities into netbox corresponding structs
- adjust unit tests
